requirements.txt
flask
opencv-python
torch
torchvision
Pillow
pandas
requests
ultralytics
flask-socketio
eventlet